Changes to import_data
Loading file is no longer a decorator, but a function.

Reason: Caching didnt work because it would receive only the decorated
function name and would cache every import the same. Changing the order of
decorators needed changes to the loading function and I moved it from a
decorator to a simple function because I think it makes the process easier
to understand